Move rfofmsg into pox/ext
