+++ Makefile
@@ -10,7 +10,7 @@ export BUILD_OBJ_DIR=$(BUILD_DIR)/obj
 export RFLIB_NAME=rflib
 
 #the lib subdirs should be done first
-export libdirs := ipc types openflow
+export libdirs := ipc types
 export srcdirs := rfclient
 
 export CPP := g++
