+++ pox/ext/rfproxy.py
@@ -10,9 +10,9 @@ import pymongo as mongo
 import rflib.ipc.IPC as IPC
 import rflib.ipc.MongoIPC as MongoIPC
 from rflib.ipc.RFProtocol import *
-from rflib.openflow.rfofmsg import *
 from rflib.ipc.RFProtocolFactory import RFProtocolFactory
 from rflib.defs import *
+from rfofmsg import *
 
 FAILURE = 0
 SUCCESS = 1
