--- rflib/openflow/Makefile
@@ -1,5 +1,0 @@
-LIBDEP=1
-
-PLIBS :=
-
-include ../../Make.rules
